backend/services/ai-services/src/tools/cost/cost_analyzer.py
"""
Cost Analyzer Tool
Provides cost analysis and breakdown functionality
"""

import asyncio
from datetime import datetime, timedelta
from typing import Dict, Any, List, Optional
import boto3
from azure.identity import DefaultAzureCredential
from azure.mgmt.costmanagement import CostManagementClient
from google.cloud import billing_v1


class CostAnalyzer:
    """Tool for analyzing cloud costs across providers"""
    
    def __init__(self):
        self.aws_client = None
        self.azure_client = None
        self.gcp_client = None
    
    async def analyze_aws_costs(self, days: int = 30) -> Dict[str, Any]:
        """Analyze AWS costs for the specified period"""
        try:
            # Placeholder implementation
            return {
                "provider": "aws",
                "period_days": days,
                "total_cost": 0.0,
                "services": {},
                "trends": {},
                "anomalies": []
            }
        except Exception as e:
            return {"error": str(e)}
    
    async def analyze_azure_costs(self, days: int = 30) -> Dict[str, Any]:
        """Analyze Azure costs for the specified period"""
        try:
            # Placeholder implementation
            return {
                "provider": "azure",
                "period_days": days,
                "total_cost": 0.0,
                "services": {},
                "trends": {},
                "anomalies": []
            }
        except Exception as e:
            return {"error": str(e)}
    
    async def analyze_gcp_costs(self, days: int = 30) -> Dict[str, Any]:
        """Analyze GCP costs for the specified period"""
        try:
            # Placeholder implementation
            return {
                "provider": "gcp",
                "period_days": days,
                "total_cost": 0.0,
                "services": {},
                "trends": {},
                "anomalies": []
            }
        except Exception as e:
            return {"error": str(e)}
    
    async def get_cost_breakdown(self, provider: str, days: int = 30) -> Dict[str, Any]:
        """Get detailed cost breakdown by service"""
        if provider.lower() == "aws":
            return await self.analyze_aws_costs(days)
        elif provider.lower() == "azure":
            return await self.analyze_azure_costs(days)
        elif provider.lower() == "gcp":
            return await self.analyze_gcp_costs(days)
        else:
            return {"error": f"Unsupported provider: {provider}"} 