backend/services/ai-services/requirements.txt
# AI Ops Guardian Angel - AI Services Requirements
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain and AI/ML frameworks
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
langchain-experimental==0.0.47
langchain-core==0.1.10

# OpenAI and Anthropic
openai==1.3.7
anthropic==0.7.8

# Database drivers
pymongo==4.6.0
motor==3.3.2
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.1

# ML and data processing
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
tensorflow==2.15.0
torch==2.1.1
transformers==4.36.2

# Cloud provider SDKs
boto3==1.34.0
azure-mgmt-compute==30.3.0
google-cloud-compute==1.14.0
google-cloud-storage==2.10.0

# Security and authentication
cryptography==41.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0
colorama==0.4.6

# Git integration
gitpython==3.1.40
pygithub==1.59.1

# Web scraping and HTTP
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2

# Background tasks
celery==5.3.4
redis==5.0.1

# Configuration and environment
python-dotenv==1.0.0
click==8.1.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
jinja2==3.1.2
markdown==3.5.1

# Optional: Jupyter for development
jupyter==1.0.0
ipython==8.18.1 